Extract FirefoxAccount's OAuth methods in their own file
